# Improve performance of CodeWriter.ToString()