--- a/src/AutoRest.CSharp/Common/Generation/Writers/CodeWriter.cs
+++ b/src/AutoRest.CSharp/Common/Generation/Writers/CodeWriter.cs
@@ -661,19 +661,15 @@ namespace AutoRest.CSharp.Generation.Writers
             }
 
             // Normalize newlines
-            //builder.Append(_builder.AsSpan(0, _length).Trim());
             var spanLines = _builder.AsSpan(0, _length).EnumerateLines();
             foreach (var line in spanLines)
             {
                 builder.Append(line.TrimEnd());
                 builder.AppendLine();
             }
+            // Remove last new line
             builder.Remove(builder.Length - Environment.NewLine.Length, Environment.NewLine.Length);
-            //builder.AppendLine(new string(_builder.AsSpan(0, _length).Trim()).Replace(_newLine, Environment.NewLine));
             return builder.ToString();
-
-            // remove any trailing whitespace, for SA1028. can roll back this change after Roslyn fixes https://github.com/dotnet/roslyn/issues/28818
-            //return string.Join(Environment.NewLine, builder.ToString().Split(Environment.NewLine).Select(l => l.TrimEnd()));
         }
 
         internal class CodeWriterScope : IDisposable