Reduces allocations in CodeWriter.ToString()

The benchmark below is the extracted performance of just ToString with 3 sample file sizes

small file
| Method    | Mean       | Error      | StdDev     | Median     | Ratio | RatioSD | Gen0     | Gen1     | Gen2     | Allocated  | Alloc Ratio |
|---------- |-----------:|-----------:|-----------:|-----------:|------:|--------:|---------:|---------:|---------:|-----------:|------------:|
| New       |   1.869 us |  0.0371 us |  0.0943 us |   1.829 us |  0.32 |    0.01 |   2.5826 |   0.1278 |        - |   21.13 KB |        0.36 |
| Original  |   5.818 us |  0.1160 us |  0.2711 us |   5.720 us |  1.00 |    0.00 |   7.2403 |   0.4196 |        - |    59.2 KB |        1.00 |

medium file
| Method    | Mean       | Error      | StdDev     | Median     | Ratio | RatioSD | Gen0     | Gen1     | Gen2     | Allocated  | Alloc Ratio |
|---------- |-----------:|-----------:|-----------:|-----------:|------:|--------:|---------:|---------:|---------:|-----------:|------------:|
| New       |  79.311 us |  0.8337 us |  0.7391 us |  79.166 us |  0.53 |    0.03 |  79.9561 |  79.9561 |  79.9561 |  267.96 KB |        0.27 |
| Original  | 145.990 us |  2.9142 us |  6.9821 us | 145.950 us |  1.00 |    0.00 | 204.1016 | 200.9277 | 162.1094 |  991.51 KB |        1.00 |

large file
| Method    | Mean       | Error      | StdDev     | Median     | Ratio | RatioSD | Gen0     | Gen1     | Gen2     | Allocated  | Alloc Ratio |
|---------- |-----------:|-----------:|-----------:|-----------:|------:|--------:|---------:|---------:|---------:|-----------:|------------:|
| New       |  83.427 us |  4.0774 us | 12.0223 us |  82.157 us |  0.23 |    0.04 | 131.9580 | 129.7607 | 129.6387 |   505.6 KB |        0.25 |
| Original  | 359.336 us | 11.4473 us | 33.2108 us | 360.003 us |  1.00 |    0.00 | 408.2031 | 408.2031 | 329.1016 | 2000.55 KB |        1.00 |